+++ game.h
@@ -6,24 +6,24 @@
 #include <sstream>
 
 class Player {
-    long _playerZone = 0; //Area of field where player is
-    long _calvinBall = 0; //=1 if player possesses Calvinball
+    int _playerZone = 0; //Area of field where player is
+    int _calvinBall = 0; //=1 if player possesses Calvinball
     std::string _score = ""; //score can be number, letter, gibberish, etc
-    long _oppPole = 0; //=1 if opposite pole used
-    long _boomerang = 0; //=1 if boomerang zone used
+    int _oppPole = 0; //=1 if opposite pole used
+    int _boomerang = 0; //=1 if boomerang zone used
 public:
     Player();
-    Player(const long sector);
+    Player(const int sector);
     
-    void setPlayerSector(long sector);
-    void setPossession(long ball); //change if player has the Calvinball or not
+    void setPlayerSector(int sector);
+    void setPossession(int ball); //change if player has the Calvinball or not
     void setOppPoleUse(); //change to 1 if player has touched the opposite pole
     void setBoomerangUse(); //change to 1 if player has called Boomerang Zone
     
-    long getPossession() const;
-    long getOppPoleUse() const;
-    long getBoomerangUse() const;
-    long getPlayerSector() const;
+    int getPossession() const;
+    int getOppPoleUse() const;
+    int getBoomerangUse() const;
+    int getPlayerSector() const;
 };
 
 void menu(); //menu text
